Added new GET method to print message by id

+++ src/main/java/com/example/greetingapp/Repository/SpringRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpringRepo extends JpaRepository<Data,Integer> {
+public interface SpringRepo extends JpaRepository<Data,Long> {
 }
